Refactor: move cart into separate folder
