+++ mypets-frontend/components/Cart/CartModalProductCard/CartModalProductCard.js
@@ -18,10 +18,10 @@ import NextImage from "next/image";
 import NextLink from "next/link";
 import { DeleteIcon } from "@chakra-ui/icons";
 
-import MerchantBadge from "../Merchant/MerchantBadge/MerchantBadge";
+import MerchantBadge from "../../Merchant/MerchantBadge/MerchantBadge";
 import CartProductQuantityPicker from "../CartProductQuantityPicker/CartProductQuantityPicker";
-import { imageToUrl } from "../../utils/urls";
-import AuthContext from "../../context/AuthContext";
+import { imageToUrl } from "../../../utils/urls";
+import AuthContext from "../../../context/AuthContext";
 
 /**
  * Product with different variations
