+++ mypets-frontend/components/Home/AnimalCategorySection/index.js
