+++ mypets-frontend/components/Layouts/Sidebar/Sidebar.js
@@ -16,9 +16,9 @@ import { useRouter } from "next/router";
 
 // import SectionHeader from "../../Layouts/SectionHeader/SectionHeader";
 import SearchbarGroup from "../../Common/SearchbarGroup/SearchbarGroup";
-import NavbarCartModalBtn from "../../NavbarCartModalBtn/NavbarCartModalBtn";
-import NavbarUserModalBtn from "../../NavbarUserModalBtn/NavbarUserModalBtn";
-import NavbarUserIcon from "../../NavbarUserIcon/NavbarUserIcon";
+import NavbarCartModalBtn from "./NavbarCartModalBtn/NavbarCartModalBtn";
+import NavbarUserModalBtn from "./NavbarUserModalBtn/NavbarUserModalBtn";
+import NavbarUserIcon from "./NavbarUserIcon/NavbarUserIcon";
 import AuthContext from "../../../context/AuthContext";
 import LoginModalBtn from "../../LoginModalBtn/LoginModalBtn";
 // import SignupModalBtn from "../../Form/SignupModalBtn/SignupModalBtn";
