+++ mypets-frontend/components/Order/OrderProductReviewModalBtn/OrderProductReviewModalBtn.js
@@ -11,11 +11,11 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 
-import MypetsBtn from "../Common/MypetsBtn/MypetsBtn";
-import Rating from "../Common/Rating/Rating";
-import ReviewTextArea from "../ReviewTextArea/ReviewTextArea";
-import { callAPI } from "../../context/AuthContext";
-import AuthContext from "../../context/AuthContext";
+import MypetsBtn from "../../Common/MypetsBtn/MypetsBtn";
+import Rating from "../../Common/Rating/Rating";
+import ReviewTextArea from "../../Common/ReviewTextArea/ReviewTextArea";
+import { callAPI } from "../../../context/AuthContext";
+import AuthContext from "../../../context/AuthContext";
 
 function OrderProductReviewModalBtn({ order_product }) {
   const { isOpen, onOpen, onClose } = useDisclosure();
